Changed the Menu colour to a grey and the back of the announcement boxes to a grey
i have changed these to see if it will define the menu and the announcements better, one thing i wanted to do but couldnt find the code for was the colour and text on the login screen i assume you didn't write the login in CSS but i could find the the colour codes for the login system because as you may know the dark theme breaks the login screen, Further changes i look at making are with the dark theme links page because the Blue on black doesn't look very professional; i also think it could be beneficial to code in a Free room and assembly schedule instead of just having it as notices. Sorry for the extended response. 
Dom.